+++ unity/Runtime/Bindings/MjBindings.cs
@@ -573,6 +573,7 @@ public enum mjtBuiltin : int{
   mjBUILTIN_GRADIENT = 1,
   mjBUILTIN_CHECKER = 2,
   mjBUILTIN_FLAT = 3,
+  mjBUILTIN_NOISE = 4,
 }
 public enum mjtMark : int{
   mjMARK_NONE = 0,